fix compilation

+++ contracts/mocks/docs/account/MyAccountMultiSignerWeighted.sol
@@ -7,7 +7,7 @@ import {Account} from "../../../account/Account.sol";
 import {EIP712} from "../../../utils/cryptography/EIP712.sol";
 import {ERC721Holder} from "../../../token/ERC721/utils/ERC721Holder.sol";
 import {ERC1155Holder} from "../../../token/ERC1155/utils/ERC1155Holder.sol";
-import {ERC7739} from "../../../utils/cryptography/signers/ERC7739.sol";
+import {ERC7739} from "../../../utils/cryptography/signers/draft-ERC7739.sol";
 import {ERC7821} from "../../../account/extensions/ERC7821.sol";
 import {Initializable} from "../../../proxy/utils/Initializable.sol";
 import {MultiSignerERC7913Weighted} from "../../../utils/cryptography/signers/MultiSignerERC7913Weighted.sol";
@@ -23,7 +23,7 @@ contract MyAccountMultiSignerWeighted is
 {
     constructor() EIP712("MyAccountMultiSignerWeighted", "1") {}
 
-    function initialize(bytes[] memory signers, uint256[] memory weights, uint256 threshold) public initializer {
+    function initialize(bytes[] memory signers, uint64[] memory weights, uint64 threshold) public initializer {
         _addSigners(signers);
         _setSignerWeights(signers, weights);
         _setThreshold(threshold);
@@ -37,11 +37,11 @@ contract MyAccountMultiSignerWeighted is
         _removeSigners(signers);
     }
 
-    function setThreshold(uint256 threshold) public onlyEntryPointOrSelf {
+    function setThreshold(uint64 threshold) public onlyEntryPointOrSelf {
         _setThreshold(threshold);
     }
 
-    function setSignerWeights(bytes[] memory signers, uint256[] memory weights) public onlyEntryPointOrSelf {
+    function setSignerWeights(bytes[] memory signers, uint64[] memory weights) public onlyEntryPointOrSelf {
         _setSignerWeights(signers, weights);
     }
